requirements.txt
dj-database-url==0.5.0
Django==2.0.4
django-tastypie==0.14.1
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
gunicorn==19.7.1
psycopg2==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.2
python-dotenv==0.8.2
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
virtualenv==15.2.0
virtualenvwrapper-win==1.2.5
whitenoise==3.3.1
